requirements.txt
flask
numpy
tensorflow-cpu
gunicorn